Add content-type only if not empty string
We encountered a problem with `*.idml` files. In the browsers the
`file.type` is the empty string `''`. Hence, Amazon is sent

```bash
------WebKitFormBoundary3aVHtVCO6rLQpIUz
Content-Disposition: form-data; name="content-type"

------WebKitFormBoundary3aVHtVCO6rLQpIUz
Content-Disposition: form-data; name="file"; filename="example.idml"
Content-Type: application/octet-stream
```

which causes S3 to interpret this as a `.zip` file. When later
downloading the example.idml from S3 one is presented with a
"example.idml.zip" file. Not setting the content-type it is the
empty string fixes this.